slider 영역 생성
Main => MainSlider 에서
슬라이더생성하기
